+++ src/DotNet.Meteor.Shared/Models/Details.cs
@@ -1,0 +1,11 @@
+
+namespace DotNet.Meteor.Shared {
+    public static class Details {
+        public const string AndroidEmulator = "Emulator";
+        public const string AndroidDevice = "Device";
+        public const string iOSSimulator = "iPhoneSimulator";
+        public const string iOSDevice = "iPhone";
+        public const string MacCatalyst = "Mac";
+        public const string Windows = "Windows";
+    }
+}
