+++ src/DotNet.Meteor.Shared/Models/Platforms.cs
@@ -1,0 +1,9 @@
+
+namespace DotNet.Meteor.Shared {
+    public static class Platforms {
+        public const string Android = "android";
+        public const string iOS = "ios";
+        public const string MacCatalyst = "maccatalyst";
+        public const string Windows = "windows";
+    }
+}
