+++ src/DotNet.Meteor.Shared/Models/Runtimes.cs
@@ -1,0 +1,10 @@
+
+namespace DotNet.Meteor.Shared {
+    public static class Runtimes {
+        public const string MacArm64 = "maccatalyst-arm64";
+        public const string MacX64 = "maccatalyst-x64";
+        public const string iOSArm64 = "ios-arm64";
+        public const string iOSimulatorX64 = "iossimulator-x64";
+        public const string Unspecified = null;
+    }
+}
