+++ src/DotNet.Meteor.Shared/Workspace/Project.cs
