+++ src/DotNet.Meteor.Shared/Workspace/WorkspaceAnalyzer.cs
